git-debimport lvm2_2.02.109-1.debian.tar.xz

+++ libdaemon/server/daemon-server.c
@@ -10,8 +10,9 @@
  * Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "daemon-shared.h"
+#include "daemon-io.h"
 #include "daemon-server.h"
+#include "daemon-log.h"
 
 #include <dlfcn.h>
 #include <errno.h>
@@ -26,7 +27,7 @@
 #include <unistd.h>
 #include <signal.h>
 
-#include <syslog.h>
+#include <syslog.h> /* FIXME. For the global closelog(). */
 
 #if 0
 /* Create a device monitoring thread. */
@@ -50,7 +51,9 @@ static void _exit_handler(int sig __attribute__((unused)))
 	_shutdown_requested = 1;
 }
 
-#ifdef linux
+#define EXIT_ALREADYRUNNING 13
+
+#ifdef __linux__
 
 #include <stddef.h>
 
@@ -69,6 +72,7 @@ static void _exit_handler(int sig __attribute__((unused)))
 #  define OOM_SCORE_ADJ_MIN (-1000)
 
 /* Systemd on-demand activation support */
+#  define SD_ACTIVATION_ENV_VAR_NAME "SD_ACTIVATION"
 #  define SD_LISTEN_PID_ENV_VAR_NAME "LISTEN_PID"
 #  define SD_LISTEN_FDS_ENV_VAR_NAME "LISTEN_FDS"
 #  define SD_LISTEN_FDS_START 3
@@ -128,7 +132,7 @@ union sockaddr_union {
 static int _handle_preloaded_socket(int fd, const char *path)
 {
 	struct stat st_fd;
-	union sockaddr_union sockaddr;
+	union sockaddr_union sockaddr = { .sa.sa_family = 0 };
 	int type = 0;
 	socklen_t len = sizeof(type);
 	size_t path_len = strlen(path);
@@ -143,7 +147,6 @@ static int _handle_preloaded_socket(int fd, const char *path)
 	    len != sizeof(type) || type != SOCK_STREAM)
 		return 0;
 
-	memset(&sockaddr, 0, sizeof(sockaddr));
 	len = sizeof(sockaddr);
 	if (getsockname(fd, &sockaddr.sa, &len) < 0 ||
 	    len < sizeof(sa_family_t) ||
@@ -164,6 +167,10 @@ static int _systemd_handover(struct daemon_state *ds)
 	unsigned long env_pid, env_listen_fds;
 	int r = 0;
 
+	/* SD_ACTIVATION must be set! */
+	if (!(e = getenv(SD_ACTIVATION_ENV_VAR_NAME)) || strcmp(e, "1"))
+		goto out;
+
 	/* LISTEN_PID must be equal to our PID! */
 	if (!(e = getenv(SD_LISTEN_PID_ENV_VAR_NAME)))
 		goto out;
@@ -172,7 +179,7 @@ static int _systemd_handover(struct daemon_state *ds)
 	env_pid = strtoul(e, &p, 10);
 	if (errno || !p || *p || env_pid <= 0 ||
 	    getpid() != (pid_t) env_pid)
-		;
+		goto out;
 
 	/* LISTEN_FDS must be 1 and the fd must be a socket! */
 	if (!(e = getenv(SD_LISTEN_FDS_ENV_VAR_NAME)))
@@ -188,6 +195,7 @@ static int _systemd_handover(struct daemon_state *ds)
 		ds->socket_fd = SD_FD_SOCKET_SERVER;
 
 out:
+	unsetenv(SD_ACTIVATION_ENV_VAR_NAME);
 	unsetenv(SD_LISTEN_PID_ENV_VAR_NAME);
 	unsetenv(SD_LISTEN_FDS_ENV_VAR_NAME);
 	return r;
@@ -198,7 +206,9 @@ out:
 static int _open_socket(daemon_state s)
 {
 	int fd = -1;
-	struct sockaddr_un sockaddr;
+	int file_created = 0;
+	struct sockaddr_un sockaddr = { .sun_family = AF_UNIX };
+	struct stat buf;
 	mode_t old_mask;
 
 	(void) dm_prepare_selinux_context(s.socket_path, S_IFSOCK);
@@ -214,17 +224,57 @@ static int _open_socket(daemon_state s)
 	/* Set Close-on-exec & non-blocking */
 	if (fcntl(fd, F_SETFD, 1))
 		fprintf(stderr, "setting CLOEXEC on socket fd %d failed: %s\n", fd, strerror(errno));
-	fcntl(fd, F_SETFL, fcntl(fd, F_GETFL, 0) | O_NONBLOCK);
+	if (fcntl(fd, F_SETFL, fcntl(fd, F_GETFL, 0) | O_NONBLOCK))
+		fprintf(stderr, "setting O_NONBLOCK on socket fd %d failed: %s\n", fd, strerror(errno));
 
 	fprintf(stderr, "[D] creating %s\n", s.socket_path);
-	memset(&sockaddr, 0, sizeof(sockaddr));
-	strcpy(sockaddr.sun_path, s.socket_path);
-	sockaddr.sun_family = AF_UNIX;
-
-	if (bind(fd, (struct sockaddr *) &sockaddr, sizeof(sockaddr))) {
-		perror("can't bind local socket.");
+	if (!dm_strncpy(sockaddr.sun_path, s.socket_path, sizeof(sockaddr.sun_path))) {
+		fprintf(stderr, "%s: daemon socket path too long.\n", s.socket_path);
 		goto error;
 	}
+
+	if (bind(fd, (struct sockaddr *) &sockaddr, sizeof(sockaddr))) {
+		if (errno != EADDRINUSE) {
+			perror("can't bind local socket");
+			goto error;
+		}
+
+		/* Socket already exists. If it's stale, remove it. */
+		if (stat(sockaddr.sun_path, &buf)) {
+			perror("stat failed");
+			goto error;
+		}
+
+		if (S_ISSOCK(buf.st_mode)) {
+			fprintf(stderr, "%s: not a socket\n", sockaddr.sun_path);
+			goto error;
+		}
+
+		if (buf.st_uid || (buf.st_mode & (S_IRWXG | S_IRWXO))) {
+			fprintf(stderr, "%s: unrecognised permissions\n", sockaddr.sun_path);
+			goto error;
+		}
+
+		if (!connect(fd, (struct sockaddr *) &sockaddr, sizeof(sockaddr))) {
+			fprintf(stderr, "Socket %s already in use\n", sockaddr.sun_path);
+			goto error;
+		}
+
+		fprintf(stderr, "removing stale socket %s\n", sockaddr.sun_path);
+
+		if (unlink(sockaddr.sun_path) && (errno != ENOENT)) {
+			perror("unlink failed");
+			goto error;
+		}
+
+		if (bind(fd, (struct sockaddr *) &sockaddr, sizeof(sockaddr))) {
+			perror("local socket bind failed after unlink");
+			goto error;
+		}
+	}
+
+	file_created = 1;
+
 	if (listen(fd, 1) != 0) {
 		perror("listen local");
 		goto error;
@@ -239,7 +289,7 @@ error:
 	if (fd >= 0) {
 		if (close(fd))
 			perror("close failed");
-		if (unlink(s.socket_path))
+		if (file_created && unlink(s.socket_path))
 			perror("unlink failed");
 		fd = -1;
 	}
@@ -252,7 +302,7 @@ static void remove_lockfile(const char *file)
 		perror("unlink failed");
 }
 
-static void _daemonise(void)
+static void _daemonise(daemon_state s)
 {
 	int child_status;
 	int fd;
@@ -287,8 +337,14 @@ static void _daemonise(void)
 		if (_shutdown_requested) /* Child has signaled it is ok - we can exit now */
 			exit(0);
 
-		/* Problem with child.  Determine what it is by exit code */
-		fprintf(stderr, "Child exited with code %d\n", WEXITSTATUS(child_status));
+		switch (WEXITSTATUS(child_status)) {
+		case EXIT_ALREADYRUNNING:
+			fprintf(stderr, "Failed to acquire lock on %s. Already running?\n", s.pidfile);
+			break;
+		default:
+			/* Problem with child.  Determine what it is by exit code */
+			fprintf(stderr, "Child exited with code %d\n", WEXITSTATUS(child_status));
+		}
 		exit(WEXITSTATUS(child_status));
 	}
 
@@ -301,7 +357,7 @@ static void _daemonise(void)
 		fd = rlim.rlim_cur;
 
 	for (--fd; fd >= 0; fd--) {
-#ifdef linux
+#ifdef __linux__
 		/* Do not close fds preloaded by systemd! */
 		if (_systemd_activation && fd == SD_FD_SOCKET_SERVER)
 			continue;
@@ -324,98 +380,88 @@ response daemon_reply_simple(const char *id, ...)
 
 	va_start(ap, id);
 
-	if (!(res.buffer = format_buffer("response", id, ap)))
+	buffer_init(&res.buffer);
+	if (!buffer_append_f(&res.buffer, "response = %s", id, NULL)) {
 		res.error = ENOMEM;
+		goto end;
+	}
+	if (!buffer_append_vf(&res.buffer, ap)) {
+		res.error = ENOMEM;
+		goto end;
+	}
 
+end:
 	va_end(ap);
-
 	return res;
-}
-
-struct thread_baton {
-	daemon_state s;
-	client_handle client;
-};
-
-static int buffer_rewrite(char **buf, const char *format, const char *string) {
-	char *old = *buf;
-	int r = dm_asprintf(buf, format, *buf, string);
-
-	dm_free(old);
-
-	return (r < 0) ? 0 : 1;
-}
-
-static int buffer_line(const char *line, void *baton) {
-	response *r = baton;
-
-	if (r->buffer) {
-		if (!buffer_rewrite(&r->buffer, "%s\n%s", line))
-			return 0;
-	} else if (dm_asprintf(&r->buffer, "%s\n", line) < 0)
-		return 0;
-
-	return 1;
 }
 
 static response builtin_handler(daemon_state s, client_handle h, request r)
 {
 	const char *rq = daemon_request_str(r, "request", "NONE");
+	response res = { .error = EPROTO };
 
 	if (!strcmp(rq, "hello")) {
 		return daemon_reply_simple("OK", "protocol = %s", s.protocol ?: "default",
-					   "version = %d", s.protocol_version, NULL);
+					   "version = %" PRId64, (int64_t) s.protocol_version, NULL);
 	}
 
-	response res = { .buffer = NULL, .error = EPROTO };
+	buffer_init(&res.buffer);
 	return res;
 }
 
-static void *client_thread(void *baton)
+static void *client_thread(void *state)
 {
-	struct thread_baton *b = baton;
+	thread_state *ts = state;
 	request req;
 	response res;
 
+	buffer_init(&req.buffer);
+
 	while (1) {
-		if (!read_buffer(b->client.socket_fd, &req.buffer))
+		if (!buffer_read(ts->client.socket_fd, &req.buffer))
 			goto fail;
 
-		req.cft = dm_config_from_string(req.buffer);
-		if (!req.cft)
-			fprintf(stderr, "error parsing request:\n %s\n", req.buffer);
+		req.cft = dm_config_from_string(req.buffer.mem);
 
-		res = builtin_handler(b->s, b->client, req);
+		if (!req.cft)
+			fprintf(stderr, "error parsing request:\n %s\n", req.buffer.mem);
+		else
+			daemon_log_cft(ts->s.log, DAEMON_LOG_WIRE, "<- ", req.cft->root);
+
+		res = builtin_handler(ts->s, ts->client, req);
 
 		if (res.error == EPROTO) /* Not a builtin, delegate to the custom handler. */
-			res = b->s.handler(b->s, b->client, req);
+			res = ts->s.handler(ts->s, ts->client, req);
 
-		if (!res.buffer) {
-			dm_config_write_node(res.cft->root, buffer_line, &res);
-			if (!buffer_rewrite(&res.buffer, "%s\n\n", NULL))
+		if (!res.buffer.mem) {
+			if (!dm_config_write_node(res.cft->root, buffer_line, &res.buffer))
+				goto fail;
+			if (!buffer_append(&res.buffer, "\n\n"))
 				goto fail;
 			dm_config_destroy(res.cft);
 		}
 
 		if (req.cft)
 			dm_config_destroy(req.cft);
-		dm_free(req.buffer);
+		buffer_destroy(&req.buffer);
 
-		write_buffer(b->client.socket_fd, res.buffer, strlen(res.buffer));
+		daemon_log_multi(ts->s.log, DAEMON_LOG_WIRE, "-> ", res.buffer.mem);
+		buffer_write(ts->client.socket_fd, &res.buffer);
 
-		free(res.buffer);
+		buffer_destroy(&res.buffer);
 	}
 fail:
 	/* TODO what should we really do here? */
-	if (close(b->client.socket_fd))
+	if (close(ts->client.socket_fd))
 		perror("close");
-	free(baton);
+	buffer_destroy(&req.buffer);
+	ts->active = 0;
 	return NULL;
 }
 
 static int handle_connect(daemon_state s)
 {
-	struct thread_baton *baton;
+	thread_state *ts;
 	struct sockaddr_un sockaddr;
 	client_handle client = { .thread_id = 0 };
 	socklen_t sl = sizeof(sockaddr);
@@ -424,51 +470,84 @@ static int handle_connect(daemon_state s)
 	if (client.socket_fd < 0)
 		return 0;
 
-	if (!(baton = malloc(sizeof(struct thread_baton))))
+	if (!(ts = dm_malloc(sizeof(thread_state)))) {
+		if (close(client.socket_fd))
+			perror("close");
+		ERROR(&s, "Failed to allocate thread state");
 		return 0;
+	}
 
-	baton->s = s;
-	baton->client = client;
+	ts->next = s.threads->next;
+	s.threads->next = ts;
 
-	if (pthread_create(&baton->client.thread_id, NULL, client_thread, baton))
+	ts->active = 1;
+	ts->s = s;
+	ts->client = client;
+
+	if (pthread_create(&ts->client.thread_id, NULL, client_thread, ts))
 		return 0;
-
-	pthread_detach(baton->client.thread_id);
 
 	return 1;
+}
+
+static void reap(daemon_state s, int waiting)
+{
+	thread_state *last = s.threads, *ts = last->next;
+	void *rv;
+
+	while (ts) {
+		if (waiting || !ts->active) {
+			if ((errno = pthread_join(ts->client.thread_id, &rv)))
+				ERROR(&s, "pthread_join failed: %s", strerror(errno));
+			last->next = ts->next;
+			dm_free(ts);
+		} else
+			last = ts;
+		ts = last->next;
+	}
 }
 
 void daemon_start(daemon_state s)
 {
 	int failed = 0;
+	log_state _log = { { 0 } };
+	thread_state _threads = { .next = NULL };
+
 	/*
 	 * Switch to C locale to avoid reading large locale-archive file used by
 	 * some glibc (on some distributions it takes over 100MB). Some daemons
 	 * need to use mlockall().
 	 */
-	if (setenv("LANG", "C", 1))
-		perror("Cannot set LANG to C");
+	if (setenv("LC_ALL", "C", 1))
+		perror("Cannot set LC_ALL to C");
 
-#ifdef linux
+#ifdef __linux__
 	_systemd_activation = _systemd_handover(&s);
 #endif
 
 	if (!s.foreground)
-		_daemonise();
+		_daemonise(s);
 
-	/* TODO logging interface should be somewhat more elaborate */
-	openlog(s.name, LOG_PID, LOG_DAEMON);
+	s.log = &_log;
+	s.log->name = s.name;
+	s.threads = &_threads;
 
-	(void) dm_prepare_selinux_context(s.pidfile, S_IFREG);
+	/* Log important things to syslog by default. */
+	daemon_log_enable(s.log, DAEMON_LOG_OUTLET_SYSLOG, DAEMON_LOG_FATAL, 1);
+	daemon_log_enable(s.log, DAEMON_LOG_OUTLET_SYSLOG, DAEMON_LOG_ERROR, 1);
 
-	/*
-	 * NB. Take care to not keep stale locks around. Best not exit(...)
-	 * after this point.
-	 */
-	if (dm_create_lockfile(s.pidfile) == 0)
-		exit(1);
+	if (s.pidfile) {
+		(void) dm_prepare_selinux_context(s.pidfile, S_IFREG);
 
-	(void) dm_prepare_selinux_context(NULL, 0);
+		/*
+		 * NB. Take care to not keep stale locks around. Best not exit(...)
+		 * after this point.
+		 */
+		if (dm_create_lockfile(s.pidfile) == 0)
+			exit(EXIT_ALREADYRUNNING);
+
+		(void) dm_prepare_selinux_context(NULL, 0);
+	}
 
 	/* Set normal exit signals to request shutdown instead of dying. */
 	signal(SIGINT, &_exit_handler);
@@ -478,10 +557,10 @@ void daemon_start(daemon_state s)
 	signal(SIGALRM, &_exit_handler);
 	signal(SIGPIPE, SIG_IGN);
 
-#ifdef linux
+#ifdef __linux__
 	/* Systemd has adjusted oom killer for us already */
 	if (s.avoid_oom && !_systemd_activation && !_protect_against_oom_killer())
-		syslog(LOG_ERR, "Failed to protect against OOM killer");
+		ERROR(&s, "Failed to protect against OOM killer");
 #endif
 
 	if (!_systemd_activation && s.socket_path) {
@@ -495,7 +574,8 @@ void daemon_start(daemon_state s)
 		kill(getppid(), SIGTERM);
 
 	if (s.daemon_init)
-		s.daemon_init(&s);
+		if (!s.daemon_init(&s))
+			failed = 1;
 
 	while (!_shutdown_requested && !failed) {
 		fd_set in;
@@ -504,20 +584,28 @@ void daemon_start(daemon_state s)
 		if (select(FD_SETSIZE, &in, NULL, NULL, NULL) < 0 && errno != EINTR)
 			perror("select error");
 		if (FD_ISSET(s.socket_fd, &in))
-			if (!handle_connect(s))
-				syslog(LOG_ERR, "Failed to handle a client connection.");
+			if (!_shutdown_requested && !handle_connect(s))
+				ERROR(&s, "Failed to handle a client connection.");
+		reap(s, 0);
 	}
 
-	if (s.socket_fd >= 0)
+	INFO(&s, "%s waiting for client threads to finish", s.name);
+	reap(s, 1);
+
+	/* If activated by systemd, do not unlink the socket - systemd takes care of that! */
+	if (!_systemd_activation && s.socket_fd >= 0)
 		if (unlink(s.socket_path))
 			perror("unlink error");
 
 	if (s.daemon_fini)
-		s.daemon_fini(&s);
+		if (!s.daemon_fini(&s))
+			failed = 1;
 
-	syslog(LOG_NOTICE, "%s shutting down", s.name);
-	closelog();
-	remove_lockfile(s.pidfile);
+	INFO(&s, "%s shutting down", s.name);
+
+	closelog(); /* FIXME */
+	if (s.pidfile)
+		remove_lockfile(s.pidfile);
 	if (failed)
 		exit(1);
 }
